pyproject.toml
[project]
name = "banco-agil"
version = "0.1.0"
description = "Sistema multi-agente inteligente para o Banco Agil"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "google-genai>=1.0",
    "chainlit>=2.0",
    "faker>=30.0",
    "python-dotenv>=1.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
